finition du player et ajout swup transition

+++ app/Http/Controllers/AlbumController.php
@@ -48,36 +48,61 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function show(Album $album)
+    public function show(Album $album, Titre $titre)
     {
         $liked = Like::where('user_id', Auth::user()->id)
                         ->where('album_id', $album->id)
                         ->first();
+
         $titres = Titre::where('album_id', $album->id)->get();
+
+        $titresId = $titres->pluck('id');
+        $likedTitres = Like::where('user_id', Auth::user()->id)
+        ->whereIn('titre_id', $titresId)
+        ->get();
         $artiste = Artiste::where('id', $album->artiste_id)->get();
 
-        return view('pages.album', compact('album', 'artiste', 'titres', 'liked'));
+        return view('pages.album', compact('album', 'artiste', 'titres', 'liked', 'likedTitres'));
     }
 
     //ajouter ou supprimer un like sur un album
     public function like(Request $request, Album $album, Like $like)
     {
-        $liked = Like::where('user_id', Auth::user()->id)
-                        ->where('album_id', $album->id)
-                        ->first();
+        if($request->has('album_id')){
+            $liked = Like::where('user_id', Auth::user()->id)
+            ->where('album_id', $album->id)
+            ->first();
 
-        $like->user_id = Auth::user()->id;
-        $like->album_id = $request->album_id;
+            $like->user_id = Auth::user()->id;
+            $like->album_id = $request->album_id;
 
-        if(!isset($liked)) {
-            $like->save();
-            $liked = true;
-        } else {
-            $liked->delete();
-            $liked = false;
+            if(!isset($liked)) {
+                $like->save();
+                $liked = true;
+            } else {
+                $liked->delete();
+                $liked = false;
+            }
+            return response()->json(['liked'=>$liked]);
         }
 
-        return response()->json(['liked'=>$liked]);
+        if($request->has('titre_id')) {
+            $likedTitre = Like::where('user_id', Auth::user()->id)
+            ->where('titre_id', $request->titre_id)
+            ->first();
+
+            $like->user_id = Auth::user()->id;
+            $like->titre_id = $request->titre_id;
+
+            if(!isset($likedTitre)) {
+                $like->save();
+                $likedTitre = true;
+            } else {
+                $likedTitre->delete();
+                $likedTitre = false;
+            }
+            return response()->json(['likedTitre'=>$likedTitre]);
+        }
     }
 
     /**
